Convert account information to editable card
